[YDB#12] [NARS] [estess] Maintain clue (to speed up future database accesses) for $order(^x("")) case and few other similar cases
Release Note
-------------
Repeated calls to $order(^xxx("")) where xxx is a global name run faster. (YDB#12)

Test
-----
* E_ALL run many times to ensure no regressions.

README
-------
The main change was to gvcst_search.c. If the search key lands before the first key in the block (a special case),
the clue was previously not being maintained. This meant future search of the same key would redo the entire
traversal of the Global Variable Tree. $order(^x("")) is an example of this kind of search where the $order function
constructs a non-existent search key that is guaranteed to be before the first key in any leaf block. Turns out all
that is needed is to invalidate the first_rec part of the clue. This way searches of keys >= clue and <= last_rec can
still be satisfied using the clue even though the first_rec is unusable. Since there was already a scheme to
invalidate the first_rec part of the clue, all that was needed was to enhance the clue maintenance to do something
similar in this special case.

In addition, there were some miscellaneous changes.

* Dbg-only changes needed to GVT_CLUE_INVALIDATE_FIRST_REC macro to take care of a few asserts
* Fixing some comments in gvcst_blk_search.c
* Add an assert in gvcst_blk_search.h
* Adjust an assert in gvcst_kill.c since the clue is now non-zero in more situations than before (triggers/testxecute
	subtest failed this assert).