[#100] Do not treat fractional timeouts with more than 3 digits after the decimal point as a 2Gi msec timeout
Basically, a READ Y:.0001 hangs indefinitely. But a READ Y:.001 times out correctly.
This is an issue introduced in GTM-5250 as part of GT.M V6.3-002.

The MV_FORCE_MSTIMEOUT macro is what converted the fractional timeout into an integer # of milliseconds.
This worked correctly if the mvtype of the mval holding the timeout value is MV_INT (which is the case
even for fractions upto 3 decimal digits). But for fractions greater than 3 decimal digits, the mvtype
would not be MV_INT but instead MV_NM and in that case, the macro was incorrectly setting the timeout
to be MAXPOSINST4 (2Gi) milliseconds (~ 24 days). This is now fixed by doing a mval2double() call to
determine the real value and converting that into milliseconds and capping that timeout with MAXPOSINT4.

The affected commands are JOB, LOCK, OPEN, READ. Additionally, WRITE /WAIT, WRITE /LISTEN, WRITE /ACCEPT
and WRITE /TLS (all of which are applicable only to the socket device type) are affected.