[#109] New ydb_repl_filter_timeout env var to control the timeout before a FILTERTIMEDOUT error is issued by the source server
Release Note
-------------
The environment variable ydb_repl_filter_timeout can be set to an integer value indicating the timeout (in seconds) that the replication source server sets for a response from the external filter program. A value less than 32 would be treated as if 32 was specified. A value greater than 131072 (2**17) would be treated as if 131072 was specified. The default value of the timeout (if env var is not specified) is 64 seconds. This provides the user a way to avoid seeing FILTERTIMEDOUT errors from the source server on relatively slower systems. YottaDB r1.10 and prior versions used to set the timeout to a hardcoded value of 64 after which a FILTERTIMEDOUT error was issued in the source server log and the external filter program was stopped.