[#122] Fix various issues with MUPIP JOURNAL ROLLBACK in handling NULL type of journal records
mur_back_process.c : It is possible a NULL record got written on behalf of a TP transaction's records in one region
(by "jnl_phase2_salvage", when a process gets abnormally killed) but some regions of the same TP transaction have
those records in tact. In this scenario, mur_back_process.c currently issues a DUPTOKEN error in pro and an assert
failure in dbg. This is now fixed to treat all records of this transaction across multiple regions (NULL or non-NULL
type of records) as belonging to the same transaction and not issue a DUPTOKEN error.

mur_forward_play_cur_jrec.c : Ensure a NULL record written on behalf of a TP transaction's records in one region
gets treated as a broken transaction and ends up in the broken transaction file).

jnlext_merge_sort_prepare.c : Use jnl seqno as token for NULL journal record (instead of 0). This ensures NULL
records get extracted in seqno order in the journal extract, broken transaction and lost transaction files.