[#97] ydb_msgprefix env var controls YottaDB message prefix for merrors.msg errors (default is "YDB")
If the env var is unset, the default prefix is "YDB"

If the env var is set to a string that is less than or equal to 31 bytes long, the env var
value is taken as the message prefix.

If the env var is set to a string that is more than 31 bytes long, the default prefix of "YDB"
is assumed.

Note the this env var only affects error messages in merrors.msg and not other the errors in
other .msg files (e.g. ydberrors.msg, gdeerrors.msg etc.)

Also the ydb_msgprefix initialization is done at the top of gtm_env_init() (instead of the
usual practice which is to add it to the end of the file) to ensure user-specified msgprefix
is honored if/when issuing errors inside gtm_env_init().