[#94] Fix usage of march=armv6 in sr_linux/gtm_env_sp.csh (used by test system) to be in sync with that in sr_linux/platform.cmake (used by cmake build)
The previous #94 commit added a "# error UNSUPPORTED PLATFORM" in sr_unix/mdefsp.h which
would kick in if the system is an ARMV6 and the __ARM_ARCH_6__ macro is not defined. This
macro is defined only in case -march=armv6 is specified (which is the case in the cmake
build). But the test system uses settings from sr_linux/gtm_env_sp.csh where -march=armv6zk
is defined which instead defines the macro __ARM_ARCH_6ZK__ and therefore reaches the
UNSUPPORTED PLATFORM codepath in mdefsp.h which compiling C programs during testing. The
fix is to keep the two in sync.