Use ydb and ydb_env_set scripts instead of gtm and gtmprofile scripts, and ydb_ prefix environment variables
When installing YottaDB, gtm is created as a symbolic link to ydb, and
gtmprofile is a symbolic link to ydb_env_set. The default names for
global directory, database files, and journal files are changed, but
if current files exist, they are Environment setting is done with an M
program rather than with shell commands.