[#83] Enhance M READ Editing in terminal to have history of 99 lines (same as direct mode)
M READ Editing is a pre-existing feature in YottaDB that is enabled with USE $PRINCIPAL:EDITING
But that maintained only a history of 1 recall item. That is, a cursor up in a READ X command
would recall the most recently entered input to a previous READ X, READ *X or READ X# command.
This change enhances that history from 1 line to 99 lines. Just like direct mode operates.

Relevant details
-----------------
* A circular buffer tt_ptr->recall_array is maintained for past terminal read inputs (99).
* The global tt_ptr->recall_index points to the current index into this array.
* READ X# (iott_readfl), READ *X (iott_rdone) and READ X (iott_read which in turn calls iott_readfl)
  commands all share the same history.
* Cursor UP goes back in history (from most recently added item back to oldest item)
* Cursor DOWN goes forward in history (from oldest item in history to newest item)
* Empty lines do not get added to history.
* Duplicate lines do not get added to history.