[#188] More code changes to ensure CALLINTCOMMIT error is correctly issued
Until now, a call-in base frame is created only if we are not already in a call-in base
frame. But since simpleAPI currently uses the same call-in base frame (SFT_CI), we have
a problem. We need to disallow a TCOMMIT done inside a call-in (ydb_ci) invocation in the
user-specified C function that is invoked inside a TP transaction created by ydb_tp_s(). We
issue a CALLINTCOMMIT error in that case but require that TREF(gtmci_nested_level) is
incremented in the ydb_ci() invocation after the ydb_tp_s() to know this is a TCOMMIT in
M code versus the op_tcommit() call done in ydb_tp_s(). We achieve this in ydb_ci_exec()
(and ydb_cij()) by calling ydb_nested_callin() if this is a ydb_ci() invocation done
inside a ydb_tp_s() invocation. And remember to unwind this call-in frame before returning
from the ydb_ci_exec() or ydb_cij() call, whether or not there was an error or success.