[#193] Fix GDE @<filename> to not report %YDB-E-IOEOF error
This is an issue in r1.20 after the ydb_msgprefix changes which caused the default
message prefix to be "YDB" instead of "GTM". This means the actual error GDE gets
when the end of <filename> is reached is YDB-E-IOEOF whereas GDE was still looking
for GTM-E-IOEOF. This is now fixed to search for -E-IOEOF since the message prefix
can be variable (depending on the ydb_msgprefix env var contents).