[#235] Remove dead debug-only code related to WBTEST_JNL_SWITCH_EXPECTED handling
The assert was needed previously to ensure the source server never invoked
jnl_file_open_switch(). But in GT.M V5.5-000, the variable "switch_and_retry"
was introduced and is set to FALSE in case of the source server so the source
server will never invoke jnl_file_open_switch(). So there is no need of the
white-box only assert. Since this is the only use of the white-box variable,
the white-box macro can also be removed and is now marked unused in wbox_test_init.h.